created src dir and restructured files out of root
