+++ apispecs/paths/get-logs.yml
@@ -3,7 +3,7 @@ get:
     - Logs
   summary: Retrieve application logs
   description: Returns the contents of the application logs.
-  operationId: logs_handler.get_logs
+  operationId: src.logs_handler.get_logs
   responses:
     '200':
       description: Successfully retrieved logs
