+++ apispecs/paths/get-one-token.yml
@@ -1,5 +1,5 @@
 get:
-  operationId: "tokens.read_one"
+  operationId: "src.tokens.read_one"
   tags:
     - App Tokens (Integration Manager Only)
   summary: "Read one token"
