+++ apispecs/security/_index.yml
@@ -1,9 +1,9 @@
 basic:
   type: 'http'
   scheme: 'basic'
-  x-basicInfoFunc: auth_handler.swag_auth
+  x-basicInfoFunc: src.auth_handler.swag_auth
 jwt:
   type: http
   scheme: bearer
   bearerFormat: JWT
-  x-bearerInfoFunc: auth_handler.decode_token
+  x-bearerInfoFunc: src.auth_handler.decode_token
