+++ build_test_db.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from config import app, db
-from models import Note, KalturaAppToken, User, AccessRestrictions, \
+from src.models import Note, KalturaAppToken, User, AccessRestrictions, \
     AppTokenSessionDefaults, UICustomizations, VendorProxies
 
 SAMPLE_TOKENS = [
